+++ Trarizon.Yieliception/AsyncYieliceptableIterator.cs
@@ -1,6 +1,6 @@
 ﻿using System.Diagnostics;
 using Trarizon.Library.Collections.Helpers;
-using Trarizon.Library.Extensions;
+using Trarizon.Library.Helpers;
 using Trarizon.Yieliception.Components;
 using Trarizon.Yieliception.Yieliceptors;